requirements.txt
# ============================================================
# Financial Data Quality Monitor — Python Dependencies
#
# NOTE: Apache Airflow is NOT listed here because it's
# pre-installed in the base Docker image. These are the
# additional packages installed on top of the Airflow image.
# ============================================================

# ── Database ─────────────────────────────────────────────────
psycopg2-binary==2.9.9
alembic>=1.13.0,<1.14

# ── Configuration ────────────────────────────────────────────
python-dotenv>=1.0.0,<2.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0

# ── Data Ingestion ───────────────────────────────────────────
yfinance>=0.2.36,<0.3
fredapi>=0.5.2,<0.6
alpha-vantage>=2.3.1,<3.0

# ── Data Processing & Analysis ───────────────────────────────
scipy>=1.11.0,<1.14

# ── Data Quality ─────────────────────────────────────────────
great-expectations>=0.18.0,<0.19

# ── Dashboard ────────────────────────────────────────────────
streamlit>=1.31.0,<1.35
plotly>=5.18.0,<6.0

# ── Market Calendar ──────────────────────────────────────────
exchange-calendars>=4.5.0,<5.0
pandas-market-calendars>=4.4.0,<5.0

# ── Retry & Logging ─────────────────────────────────────────
tenacity>=8.2.0,<9.0
structlog>=24.1.0,<25.0